Programming Rust - Chapter 2 - Pt 1
Scripts developed for the beggining section of the second chapter of the book "Programming Rust". Chapter 2, called "A Tour of Rust" aims to show the main things that differentiate Rust from other languages. Here we cover the examples that showcase:
- Functions (Greatest Common Divisor Function)
- Tests (GCD Function)
- A small web server (Actix Web)
- Parallel computing (Plotting of the mandelbrot set).

This last section still lacks a main function, as I haven't reached that part in the book. This is up to page 55.